requirements.txt
PyYAML==6.0.2
requests==2.26.0
